requirements.txt
requests>=2.25.0
pillow>=8.0.0
numpy>=1.19.0
torch>=1.9.0
torchvision>=0.10.0 